fixed api bugus

+++ client/src/pages/Customers.js
@@ -63,13 +63,21 @@ const Customers = () => {
 
   const handleDeleteCustomer = async (customerId) => {
     try {
+      // Display a confirmation dialog
+      const isConfirmed = window.confirm("Are you sure you want to delete this customer?");
+  
+      if (!isConfirmed) {
+        // User canceled the deletion
+        return;
+      }
+  
       const res = await axios.delete(
         `http://localhost:5000/api/v1/customers/${customerId}`, {
-            headers: {
-              Authorization: `Bearer ${token}`,
-            },
-          });
-
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
+        });
+  
       if (res.data.success) {
         // Reload the customers after deletion
         setCurrentPage(1);
@@ -79,6 +87,7 @@ const Customers = () => {
       console.error("Error deleting customer:", error);
     }
   };
+  
 
   return (
     <div className="flex-1 p-4 md:p-8 pt-6 mt-10 sm:ml-48 justify-center items-center overflow-y-auto max-h-screen">
